team15/pyproject.toml
#
#     /\_____/\
#    /  o   o  \
#   ( ==  ^  == )
#    )         (
#   (           )
#  ( (  )   (  ) )
# (__(__)___(__)__)
#
# References:
#   - pyproject.toml: https://packaging.python.org/en/latest/
#   - poetry        : https://python-poetry.org/docs/
#

[project]
name            = "mon"
version         = "2.7.1"
authors         = [
    {name="Long H. Pham", email="longpham3105@gmail.com"},
]
maintainers     = [
    {name="Long H. Pham", email="longpham3105@gmail.com"},
]
description     = "Just a research framework."
readme          = "README.md"
license         = "MIT"
license-files   = ["LICENSE"]
keywords        = ["computer vision", "image processing", "deep learning", "research"]
classifiers     = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
]
requires-python = ">=3.10"
dependencies    = [
    # Python Functionality
    "emoji",                            # Emoji for Python.
    "fjson",                            # Python JSON writer with formatted floats
    "loguru",                           # Python logging made (stupidly) simple.
    "plum-dispatch",                    # Multiple dispatch in Python.
    "psutil",                           # Cross-platform lib for process and system monitoring in Python.
    "pyhumps",                          # Convert strings between snake case, camel case and pascal case in Python.
    "pytest",                           # Simple powerful testing with Python.
    "python-box",                       # Advanced Python dictionaries with dot notation access.
    "requests",                         # Requests is an elegant and simple HTTP library for Python, built with ♥.
    "rich",                             # Render rich text, tables, progress bars, syntax highlighting, markdown and more to the terminal.
    "tabulate",                         # Pretty-print tabular data in Python, a library and a command-line utility.
    "tqdm",                             # Fast, Extensible Progress Meter.
    "typing-extensions",                # Backported and Experimental Type Hints for Python 3.7+.
    "validators",                       # Python Data Validation for Humans.
    # DSA
    "PyYAML",                           # YAML parser and emitter for Python.
    "filterpy",                         # Kalman filters and other optimal and non-optimal estimation filters in Python.
    "numpy",                            # The fundamental package for scientific computing with Python.
    "protobuf",                         # Protocol Buffers - Google's data interchange format.
    "scipy",                            # Scientific Library for Python.
    "xmltodict",                        # Makes working with XML feel like you are working with JSON.
    # AI
    "calflops",                         # A library for calculating the number of floating point operations (FLOPs) in a PyTorch model.
    "einops",                           # A new flavor of deep learning operations.
    "fvcore",                           # Collection of common code shared among different research projects in FAIR computer vision team.
    "lightning>=2.5.2",                 # The Deep Learning framework to train, deploy, and ship AI products Lightning fast.
    "onnx",                             # Open Neural Network Exchange.
    "onnxruntime",                      # ONNX Runtime is a performance-focused scoring engine for Open Neural Network Exchange (ONNX) models.
    "openvino",                         # OpenVINO™ is a toolkit for optimizing and deploying AI inference.
    "scikit-learn",                     # A set of python modules for machine learning and data mining.
    "snntorch",
    "spikingjelly",
    "tensorboard",                      # TensorFlow's Visualization Toolkit.
    "thop",                             # Count the MACs / FLOPs of your PyTorch model.
    "torch-fidelity",
    "torch>=2.7.1",                     # PyTorch is an optimized tensor library for deep learning using GPUs and CPUs.
    "torchfile",
    "torchmetrics>=1.7.3",              # Collection of PyTorch native metrics for easy evaluating machine learning models.
    "transformers",                     # 🤗 Transformers: State-of-the-art Machine Learning for Pytorch, TensorFlow, and JAX.
    # Vision
    "Pillow",                           # Python Imaging Library (Fork).
    "PyTurboJPEG",                      # A Python wrapper of libjpeg-turbo for decoding and encoding JPEG image.
    "albumentations",                   # An efficient library for image augmentation, providing extensive transformations to support machine learning and computer vision tasks.
    "faster-coco-eval",                 # Faster interpretation of the original COCOEval
    "ffmpeg-python",                    # Python bindings for FFmpeg—with complex filtering support.
    "kornia",                           # Kornia is a differentiable computer vision library for PyTorch.
    "opencv-contrib-python",            # Wrapper package for OpenCV python bindings.
    "opencv-python",                    # Computer vision and machine learning software library.
    "pillow_heif",                      # Python interface for libheif library.
    "pycocotools",                      # COCO dataset and evaluation metrics for object detection, instance segmentation, and person keypoint detection.
    "pyiqa",                            # PyTorch Toolbox for Image Quality Assessment.
    "rawpy",                            # rawpy is an easy-to-use Python wrapper for the LibRaw library. It also contains some extra functionality for finding and repairing hot/dead pixels.
    "scikit-image",                     # Image processing in Python.
    "torchvision",                      # Image and video datasets and models for Torch deep learning.
    "ultralytics",
    # Temporary Fix
    "bitsandbytes==0.42.0"              # Fix for MacOS: 0.45.5 error
]
# dynamic         = ["dependencies"]

[project.urls]
Homepage      = "https://github.com/phlong3105/mon"
Documentation = "https://github.com/phlong3105/mon/docs"
Repository    = "https://github.com/phlong3105/mon.git"
Issues        = "https://github.com/phlong3105/mon/issues"

[project.optional-dependencies]
docs = [
    "markdown",                         # Python implementation of Markdown.
    "mkdocs",                           # Project documentation with Markdown.
    "mkdocs-material",                  # A Material Design theme for mkdocs.
    "mkdocstrings",                     # Automatic documentation from sources, for MkDocs.
    "sphinx",                           # Sphinx is a tool that makes it easy to create intelligent and beautiful documentation.
    "sphinx-paramlinks",                # A Sphinx extension which allows :param: directives within Python documentation to be linkable.
]
gui  = [
    "PyQt6",                            # Python bindings for the Qt cross-platform application toolkit.
    "matplotlib",                       # Publication quality figures in Python.
]

[tool.poetry]
packages = [
    {include="mon", from="src"},
]
include  = []
exclude  = []

[tool.poetry.dependencies]
# snntorch      = {path="src/mon/extra/snntorch",     develop=true}
# spikingjelly  = {path="src/mon/extra/spikingjelly", develop=true}
ultralytics   = {path="src/mon/extra/ultralytics",  develop=true}

[build-system]
requires      = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
